refactor: useValidate 유틸함수로 변경
