+++ src/pages/Services/Room/components/DetailForm.tsx
@@ -7,7 +7,7 @@ import { FormInstance } from 'antd/es/form/Form';
 import * as S from '../RoomDetail.style';
 
 export default function DetailForm({ form }: { form: FormInstance }) {
-  const { required, max, pattern } = CustomForm.useValidate();
+  const { required, max, pattern } = CustomForm.validateUtils();
 
   return (
     <>
